fix: correctly renamed toast to Toast
